% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot.R
\name{aesReplacer}
\alias{aesReplacer}
\title{set aes for vector}
\usage{
aesReplacer(x, dict, type = "color")
}
\arguments{
\item{x}{a vector to be replaced with aes}

\item{dict}{a vector of colors/shapes named with categories of x.}

\item{type}{aes type, allowed: color, fill, shape.}
}
\value{
a vector of aes
}
\description{
set aes for vector
}
\examples{

}
