% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot.R
\name{checkAes}
\alias{checkAes}
\title{check aes}
\usage{
checkAes(x, df, name = NULL)
}
\arguments{
\item{x}{a value or a colname}

\item{df}{a data.frame}

\item{name}{report name}
}
\value{
a vector of nrow(df)
}
\description{
a value or a colname
}
\examples{

}
