% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot.R
\name{columnCross}
\alias{columnCross}
\title{Visualization for links between two types of nodes with annotaion plots}
\usage{
columnCross(
  edges,
  columns,
  height = 1,
  flank_mult = rep(0.1, length(columns)),
  segment_shrink = 0.1
)
}
\arguments{
\item{edges}{data.frame with at least two columns named with "source" and "target"}

\item{columns}{a list of multiple columns of nodes on, the order will be used as level.}

\item{height}{plot height, used for control coordinates}

\item{flank_mult}{mulitply with `scale`, flankings at four sides,}

\item{segment_shrink}{text may overlaid on segement, use this to avoid it}
}
\value{
a ggplot object
}
\description{
Visualization for links between two types of nodes with annotaion plots
}
\examples{

}
