% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot.R
\name{is.color}
\alias{is.color}
\title{check the vector is/are color(s) or not}
\source{
https://stackoverflow.com/a/13290832
}
\usage{
is.color(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a named vector
}
\description{
check the vector is/are color(s) or not
}
\examples{

\dontrun{
is.color(c(NA, "black", "blackk", "1", "#00", "#000000"))
}
}
