% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot.R
\name{newCoord}
\alias{newCoord}
\title{generate 1-d coord with scale and flankings}
\usage{
newCoord(x, scale = 1, flank_mult = 0.1)
}
\arguments{
\item{x}{a vector of numeric or factor}

\item{scale}{the max coord, return will be in the range of 0 and `scale`}

\item{flank_mult}{mulitply with `scale`, flankings at both sides,}
}
\value{
new 1-d coordinate
}
\description{
generate 1-d coord with scale and flankings
}
\examples{
newCoord(1:10, scale = 110, flank_mult = 0.1)

}
