% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossplot.R
\name{squareCross}
\alias{squareCross}
\title{Visualization for links between four types of nodes with annotation plots}
\usage{
squareCross(
  edges,
  top,
  bottom,
  left,
  right,
  width = 1,
  height = 1,
  show_node_label = F,
  flank_mult = c(top = 0.1, bottom = 0.1, left = 0.1, right = 0.1),
  expand_xlim_mult = c(0.2, 0.2),
  expand_ylim_mult = c(0.2, 0.2)
)
}
\arguments{
\item{edges}{data.frame with at least two columsn with nomes of "source" and "target"}

\item{top}{a vector of nodes on top, the order will be used as level.}

\item{bottom}{similar to top but on bottom}

\item{left}{similar to top but on left}

\item{right}{similar to top but on right}

\item{width}{plot width, used for control coordinates}

\item{height}{plot height, used for control coordinates}

\item{show_node_label}{whether to show node label}

\item{flank_mult}{mulitply with `scale`, flankings at four sides,}

\item{expand_xlim_mult}{expand x limits with expand_xlim_mult*width. This may be used when show_node_label is TRUE.}

\item{expand_ylim_mult}{expand y limits with expand_ylim_mult*width}
}
\value{
a ggplot object
}
\description{
Visualization for links between four types of nodes with annotation plots
}
\examples{

}
